fix signature

+++ java/vortex-jni/src/main/java/dev/vortex/jni/JNIArray.java
@@ -18,6 +18,7 @@ package dev.vortex.jni;
 import com.google.common.base.Preconditions;
 import dev.vortex.api.Array;
 import dev.vortex.api.DType;
+import java.math.BigDecimal;
 import java.util.OptionalLong;
 import org.apache.arrow.c.ArrowArray;
 import org.apache.arrow.c.ArrowSchema;
@@ -131,7 +132,7 @@ public final class JNIArray implements Array {
     }
 
     @Override
-    public double getBigDecimal(int index) {
+    public BigDecimal getBigDecimal(int index) {
         return NativeArrayMethods.getBigDecimal(pointer.getAsLong(), index);
     }
